module_builtin_ast optimized. profiler module removed from the main daScript
