+++ src/hal/project_specific.cpp
@@ -8,7 +8,6 @@ void require_project_specific_modules() {
     #if defined(_EMSCRIPTEN_VER)
     return;
     #else
-    NEED_MODULE(Module_TestProfile);
     NEED_MODULE(Module_UnitTest);
     #endif
 }
